refactor sync analyze prebuilt from url tests
