+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_invoice_from_url.py
@@ -11,13 +11,12 @@ from azure.core.exceptions import HttpResponseError
 from azure.core.credentials import AzureKeyCredential
 from azure.ai.formrecognizer._generated.v2_1.models import AnalyzeOperationResult
 from azure.ai.formrecognizer._response_handlers import prepare_prebuilt_models
-from azure.ai.formrecognizer import FormRecognizerClient, FormRecognizerApiVersion, DocumentAnalysisClient
+from azure.ai.formrecognizer import FormRecognizerClient, FormRecognizerApiVersion
 from testcase import FormRecognizerTest
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 from preparers import FormRecognizerPreparer
 
 FormRecognizerClientPreparer = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
-DocumentAnalysisClientPreparer = functools.partial(_GlobalClientPreparer, DocumentAnalysisClient)
 
 
 class TestInvoiceFromUrl(FormRecognizerTest):
@@ -79,26 +78,6 @@ class TestInvoiceFromUrl(FormRecognizerTest):
             self.assertFormFieldsTransformCorrect(invoice.fields, document_result.fields, read_results)
 
         self.assertFormPagesTransformCorrect(returned_model.pages, read_results, page_results)
-
-    @FormRecognizerPreparer()
-    @DocumentAnalysisClientPreparer()
-    def test_invoice_tiff(self, client):
-        poller = client.begin_analyze_document_from_url(model="prebuilt-invoice", document_url=self.invoice_url_tiff)
-
-        result = poller.result()
-        assert len(result.documents) == 1
-        invoice = result.documents[0]
-
-        # check dict values
-        assert invoice.fields.get("VendorName").value ==  "Contoso"
-        assert invoice.fields.get("VendorAddress").value, '1 Redmond way Suite 6000 Redmond ==  WA 99243'
-        assert invoice.fields.get("CustomerAddressRecipient").value ==  "Microsoft"
-        assert invoice.fields.get("CustomerAddress").value, '1020 Enterprise Way Sunnayvale ==  CA 87659'
-        assert invoice.fields.get("CustomerName").value ==  "Microsoft"
-        assert invoice.fields.get("InvoiceId").value ==  '34278587'
-        assert invoice.fields.get("InvoiceDate").value, date(2017, 6 ==  18)
-        assert invoice.fields.get("Items").value[0].value["Amount"].value ==  56651.49
-        assert invoice.fields.get("DueDate").value, date(2017, 6 ==  24)
 
     @pytest.mark.live_test_only
     @FormRecognizerPreparer()
