+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_receipt_from_url.py
@@ -7,110 +7,16 @@
 import pytest
 import functools
 from datetime import date, time
-from azure.core.exceptions import HttpResponseError, ServiceRequestError, ClientAuthenticationError
-from azure.core.credentials import AzureKeyCredential
 from azure.ai.formrecognizer._generated.v2_1.models import AnalyzeOperationResult
-from azure.ai.formrecognizer._generated.v2021_09_30_preview.models import AnalyzeResultOperation
 from azure.ai.formrecognizer._response_handlers import prepare_prebuilt_models
-from azure.ai.formrecognizer import FormRecognizerClient, FormRecognizerApiVersion, DocumentAnalysisClient, AnalyzeResult
+from azure.ai.formrecognizer import FormRecognizerClient, FormRecognizerApiVersion
 from testcase import FormRecognizerTest
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 from preparers import FormRecognizerPreparer
 
-GlobalClientPreparer = functools.partial(_GlobalClientPreparer, DocumentAnalysisClient)
 GlobalClientPreparerV2 = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
 
 class TestReceiptFromUrl(FormRecognizerTest):
-
-    @FormRecognizerPreparer()
-    def test_polling_interval(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        client = DocumentAnalysisClient(formrecognizer_test_endpoint, AzureKeyCredential(formrecognizer_test_api_key), polling_interval=7)
-        assert client._client._config.polling_interval ==  7
-
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg, polling_interval=6)
-        poller.wait()
-        assert poller._polling_method._timeout ==  6
-        poller2 = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg)
-        poller2.wait()
-        assert poller2._polling_method._timeout ==  7  # goes back to client default
-
-    @pytest.mark.live_test_only
-    def test_active_directory_auth(self):
-        token = self.generate_oauth_token()
-        endpoint = self.get_oauth_endpoint()
-        client = DocumentAnalysisClient(endpoint, token)
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg)
-        result = poller.result()
-        assert result is not None
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipts_encoded_url(self, client):
-        try:
-            poller = client.begin_analyze_document_from_url("prebuilt-receipt", "https://fakeuri.com/blank%20space")
-        except HttpResponseError as e:
-            assert "https://fakeuri.com/blank%20space" in  e.response.request.body
-
-    @FormRecognizerPreparer()
-    def test_receipt_url_bad_endpoint(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        with self.assertRaises(ServiceRequestError):
-            client = DocumentAnalysisClient("http://notreal.azure.com", AzureKeyCredential(formrecognizer_test_api_key))
-            poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg)
-
-    @FormRecognizerPreparer()
-    def test_receipt_url_auth_bad_key(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        client = DocumentAnalysisClient(formrecognizer_test_endpoint, AzureKeyCredential("xxxx"))
-        with self.assertRaises(ClientAuthenticationError):
-            poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_bad_url(self, client):
-        with self.assertRaises(HttpResponseError):
-            poller = client.begin_analyze_document_from_url("prebuilt-receipt", "https://badurl.jpg")
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_url_pass_stream(self, client):
-        with open(self.receipt_png, "rb") as receipt:
-            with self.assertRaises(HttpResponseError):
-                poller = client.begin_analyze_document_from_url("prebuilt-receipt", receipt)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_url_transform_jpg(self, client):
-        responses = []
-
-        def callback(raw_response, _, headers):
-            analyze_result = client._deserialize(AnalyzeResultOperation, raw_response)
-            extracted_receipt = AnalyzeResult._from_generated(analyze_result.analyze_result)
-            responses.append(analyze_result)
-            responses.append(extracted_receipt)
-
-        poller = client.begin_analyze_document_from_url(
-            "prebuilt-receipt",
-            document_url=self.receipt_url_jpg,
-            cls=callback
-        )
-
-        result = poller.result()
-        raw_analyze_result = responses[0].analyze_result
-        returned_model = responses[1]
-
-        # Check AnalyzeResult
-        assert returned_model.model_id == raw_analyze_result.model_id
-        assert returned_model.api_version == raw_analyze_result.api_version
-        assert returned_model.content == raw_analyze_result.content
-
-        self.assertDocumentPagesTransformCorrect(returned_model.pages, raw_analyze_result.pages)
-        self.assertDocumentTransformCorrect(returned_model.documents, raw_analyze_result.documents)
-        self.assertDocumentTablesTransformCorrect(returned_model.tables, raw_analyze_result.tables)
-        self.assertDocumentKeyValuePairsTransformCorrect(returned_model.key_value_pairs, raw_analyze_result.key_value_pairs)
-        self.assertDocumentEntitiesTransformCorrect(returned_model.entities, raw_analyze_result.entities)
-        self.assertDocumentStylesTransformCorrect(returned_model.styles, raw_analyze_result.styles)
-
-        # check page range
-        assert len(raw_analyze_result.pages) == len(returned_model.pages)
 
     @FormRecognizerPreparer()
     @GlobalClientPreparerV2()
@@ -184,137 +90,8 @@ class TestReceiptFromUrl(FormRecognizerTest):
         self.assertReceiptItemsHasValues(receipt.fields["Items"].value, receipt.page_range.first_page_number, True)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_url_png(self, client):
-
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_png)
-
-        result = poller.result()
-        assert len(result.documents) == 1
-        receipt = result.documents[0]
-        assert receipt.fields.get("MerchantAddress").value, '123 Main Street Redmond ==  WA 98052'
-        assert receipt.fields.get("MerchantName").value ==  'Contoso'
-        assert receipt.fields.get("Subtotal").value ==  1098.99
-        assert receipt.fields.get("Tax").value ==  104.4
-        assert receipt.fields.get("Total").value ==  1203.39
-        assert receipt.fields.get("TransactionDate").value == date(year=2019, month=6, day=10)
-        assert receipt.fields.get("TransactionTime").value == time(hour=13, minute=59, second=0)
-        receipt_type = receipt.fields.get("ReceiptType")
-        assert receipt_type.confidence is not None
-        assert receipt_type.value ==  'Itemized'
-
-        assert len(result.pages) == 1
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_multipage_url(self, client):
-
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.multipage_receipt_url_pdf)
-        result = poller.result()
-
-        assert len(result.documents) == 2
-        receipt = result.documents[0]
-        assert receipt.fields.get("MerchantAddress").value, '123 Main Street Redmond ==  WA 98052'
-        assert receipt.fields.get("MerchantName").value ==  'Contoso'
-        assert receipt.fields.get("MerchantPhoneNumber").value ==  '+19876543210'
-        assert receipt.fields.get("Subtotal").value ==  11.7
-        assert receipt.fields.get("Tax").value ==  1.17
-        assert receipt.fields.get("Tip").value ==  1.63
-        assert receipt.fields.get("Total").value ==  14.5
-        assert receipt.fields.get("TransactionDate").value == date(year=2019, month=6, day=10)
-        assert receipt.fields.get("TransactionTime").value == time(hour=13, minute=59, second=0)
-        receipt_type = receipt.fields.get("ReceiptType")
-        assert receipt_type.confidence is not None
-        assert receipt_type.value ==  'Itemized'
-        receipt = result.documents[1]
-        assert receipt.fields.get("MerchantAddress").value, '123 Main Street Redmond ==  WA 98052'
-        assert receipt.fields.get("MerchantName").value ==  'Contoso'
-        assert receipt.fields.get("Subtotal").value ==  1098.99
-        assert receipt.fields.get("Tax").value ==  104.4
-        assert receipt.fields.get("Total").value ==  1203.39
-        assert receipt.fields.get("TransactionDate").value == date(year=2019, month=6, day=10)
-        assert receipt.fields.get("TransactionTime").value == time(hour=13, minute=59, second=0)
-        receipt_type = receipt.fields.get("ReceiptType")
-        assert receipt_type.confidence is not None
-        assert receipt_type.value ==  'Itemized'
-
-        assert len(result.pages) == 2
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_multipage_transform_url(self, client):
-
-        responses = []
-
-        def callback(raw_response, _, headers):
-            analyze_result = client._deserialize(AnalyzeResultOperation, raw_response)
-            extracted_receipt = AnalyzeResult._from_generated(analyze_result.analyze_result)
-            responses.append(analyze_result)
-            responses.append(extracted_receipt)
-
-        poller = client.begin_analyze_document_from_url(
-            "prebuilt-receipt",
-            self.multipage_receipt_url_pdf,
-            cls=callback
-        )
-
-        result = poller.result()
-        raw_analyze_result = responses[0].analyze_result
-        returned_model = responses[1]
-
-        # Check AnalyzeResult
-        assert returned_model.model_id == raw_analyze_result.model_id
-        assert returned_model.api_version == raw_analyze_result.api_version
-        assert returned_model.content == raw_analyze_result.content
-        
-        self.assertDocumentPagesTransformCorrect(returned_model.pages, raw_analyze_result.pages)
-        self.assertDocumentTransformCorrect(returned_model.documents, raw_analyze_result.documents)
-        self.assertDocumentTablesTransformCorrect(returned_model.tables, raw_analyze_result.tables)
-        self.assertDocumentKeyValuePairsTransformCorrect(returned_model.key_value_pairs, raw_analyze_result.key_value_pairs)
-        self.assertDocumentEntitiesTransformCorrect(returned_model.entities, raw_analyze_result.entities)
-        self.assertDocumentStylesTransformCorrect(returned_model.styles, raw_analyze_result.styles)
-
-        # check page range
-        assert len(raw_analyze_result.pages) == len(returned_model.pages)
-
-    @pytest.mark.live_test_only
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_continuation_token(self, client):
-
-        initial_poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg)
-        cont_token = initial_poller.continuation_token()
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", None, continuation_token=cont_token)
-        result = poller.result()
-        assert result is not None
-        initial_poller.wait()  # necessary so azure-devtools doesn't throw assertion error
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_locale_specified(self, client):
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg, locale="en-IN")
-        assert 'en-IN' == poller._polling_method._initial_response.http_response.request.query['locale']
-        result = poller.result()
-        assert result
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_locale_error(self, client):
-        with pytest.raises(HttpResponseError) as e:
-            client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg, locale="not a locale")
-        assert "InvalidArgument" == e.value.error.code
-
-    @FormRecognizerPreparer()
     @GlobalClientPreparerV2(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
     def test_receipt_locale_v2(self, client):
         with pytest.raises(ValueError) as e:
             client.begin_recognize_receipts_from_url(self.receipt_url_jpg, locale="en-US")
         assert "'locale' is only available for API version V2_1 and up" in str(e.value)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_pages_kwarg_specified(self, client):
-        poller = client.begin_analyze_document_from_url("prebuilt-receipt", self.receipt_url_jpg, pages="1")
-        assert '1' == poller._polling_method._initial_response.http_response.request.query['pages']
-        result = poller.result()
-        assert result
